Fix old parents/child linking in mergePartitions

+++ viswidget.cpp
@@ -191,7 +191,9 @@ void VisWidget::drawTimescale(QPainter * painter, unsigned long long start, unsi
     // We want a round number
     unsigned long long tick_span = span / max_ticks; // Not round
     int power = floor(log10(span / max_ticks)); // How many zeros
-    int roundfactor = pow(10, power);
+    unsigned long long roundfactor = pow(10, power);
+    //std::cout << "span " << span << " max_ticks " << max_ticks << " tick_span " << tick_span << std::endl;
+    //std::cout << "power " << power << " roundfactor " << roundfactor << std::endl;
     tick_span = (tick_span / roundfactor) * roundfactor; // Now round
 
     // Now we must find the first number after startTime divisible by
